requirements.txt
psycopg2-binary==2.8.6
django-allauth==0.42.0
django==3.1.14
djangorestframework==3.12.1
pyusb==1.2.1